swupdate exit code: return exit status from main child process on main process
child processes such as start_download take care of returning a valid exit
code, but that code is lost right now.
Instead it's possible to just have swupdate forward it out of the main
process (currently exiting with success all the time),
so e.g. swupdate -d with a failing update fails appropriately