Atmark: lower verbosity for handlers found messages
list of handlers can change when using lua so upstream does not want to
put these logs down to TRACE, but it makes more sense for us as we still
use swupdate -i/-d quite a bit.
If/when we start using swupdate_client this commit can be dropped as the
logs are only printed on swupdate start, not client connecting.

Also do the same for default bootloader: we only compile with none, so
no other bootloader can possibly have been selected..