Atmark: channel_curl: fetch new URL on RANGE_ERROR with Armadillo Twin
SWU download from AWS S3 may be retried when SWUpdate is executed
during communication instability, e.g. when using LTE Cat.M1.
A signed URL with an expiry date is received in advance from AWS S3,
but if the signature expires when the download is retried, the
download will fail.

This modification adds a process that calls an external script that
re-issues the signed URL when the URL expires and changes the URL to
download.

Redmine: 63374
Change-Id: Ic7dfd60ea9014e3d76c9051a56f1c57f528d4865